Generic Bootstrapping Subroutine

Provide a numeric or character vector and create bootstrap samples (with estimates) of the original vector. Also calculates CI95 using the quantile() function.

bootstrap(x, boot = 1000, FUN, up = 0.975, lo = 0.025)

Arguments

x
Character or Numeric. The original data to be bootstrapped.

boot
Number of bootstraps to perform.

FUN
Function desired for the point estimate of the original data vector (if numeric data).

up
Upper confidence limit.

lo
Lower confidence limit.

Value

A list containing:

BootSamples

List of the bootstrap populations created during the simulation.

BootEstimates

List of the various point estimates of each of the bootstrap samples.

CI95

Vector of the point estimate and upper & lower CI95 produced via the bootstrap samples.

SE

The standard error based on the original data. Could be used to calculate CI95 via 1.96*SE if so desired.

See also

Examples

bootstrap(x = round(runif(25,1,100)), boot = 50, FUN = mean) # numeric

#> $BootSamples
#> $BootSamples[[1]]
#>  [1] 73 44 44  8 17 38 20 45  5  4 45 47 17 38 47 73 81 75 90 20 73 38  5 45 17
#> 
#> $BootSamples[[2]]
#>  [1] 73 90 52  8 17 75 45 11 47  8 90 17 90 75 58 92 58 43 44 12 73 56  5 75 81
#> 
#> $BootSamples[[3]]
#>  [1] 90 58 81 90 47 90 68 43 11 58 56 12 52 83 90 12  5 92 52 52 44 20 90 73 38
#> 
#> $BootSamples[[4]]
#>  [1]  5 47 90 11 83 47 90 20 83 38 73 44 56 38 17  8 38 56 56 92 90 94 43 92 56
#> 
#> $BootSamples[[5]]
#>  [1]  8 75 38 58 68 43 90 20 68 17 43  8 43 35 90  4 45 44 17 38 83 90 92 83  8
#> 
#> $BootSamples[[6]]
#>  [1] 56 90  4 90 35 47 20 35  8 12 38 90  5 52 12 52 56 45 12 68 52 38 44 90 11
#> 
#> $BootSamples[[7]]
#>  [1]  4 73 17  5 83 12 73 92 12 45 90  8  8 45 35 90 73 12 11 83 92 58 38 81 58
#> 
#> $BootSamples[[8]]
#>  [1] 45 75 75 83 81 90 38 83 12 44 43 94 90 43 35  5 11 35 38 90 68 58  4 83 68
#> 
#> $BootSamples[[9]]
#>  [1] 90 75 90 90 17 47 11  4 83 73 11 35 52 73 11 35 43 73 52 75 43 68 11 11 73
#> 
#> $BootSamples[[10]]
#>  [1] 17 56 11  8 47 94 75 20 45 73  8 83 17 56 12 90 52 83 11 92 90 17 52 94 92
#> 
#> $BootSamples[[11]]
#>  [1] 92 20 73 81 52 81 20  4 68  4 73 56 52  4  4 43 20 73 73 94 90 90 90 45 11
#> 
#> $BootSamples[[12]]
#>  [1] 75 90  5 12 45 68  5 45 73 38  5 20 73 20 38 90 73 47 68 83 90 12  5  4 20
#> 
#> $BootSamples[[13]]
#>  [1] 94  5 68  8 17 90 44  4 52 90 92 38 43 47 81 38 45 56 45 68 90 58 20 81 38
#> 
#> $BootSamples[[14]]
#>  [1] 92 56 35 47 38 56 81 20 52 92 81  4 94 56 52 68 11 56 56 38 45 58 52 56 58
#> 
#> $BootSamples[[15]]
#>  [1] 35 68 52  8 73 90 90 68 45 35 83 47 83 83 81 52 17 17 44 90 17 73 20 94 90
#> 
#> $BootSamples[[16]]
#>  [1] 81 35 56 90 52 90  4 94  8 73 92 94 43 12 12 47 73  5 20 44 81 43  4 68 56
#> 
#> $BootSamples[[17]]
#>  [1] 17 94 92 92 17 45 68 90 43 43 20 58 90 58 83 90 38 92 43 44 81 44 83  8 81
#> 
#> $BootSamples[[18]]
#>  [1] 90  8 45 56 56 68 52 12  4 44 58 56  8 17 94 94 83 68 90 43 44 52 94 47 73
#> 
#> $BootSamples[[19]]
#>  [1] 17 47 17 73  8 68 90 58  4 45 43 68 45 44 73 81 52 47 73 12 73 90 90 75 20
#> 
#> $BootSamples[[20]]
#>  [1] 73 38 52 38 44 52 73 92 52 11 81  8 35 45 35 47 12 11 44 35 58 20 68 47  4
#> 
#> $BootSamples[[21]]
#>  [1]  8 75 35 75 58 35 90 20 38 52 56 92 94  8 81 83 47 68  8 52 38 73 12 92 45
#> 
#> $BootSamples[[22]]
#>  [1] 47 17 73 52 35 20 92 44 58 92 94 81 45 38 35 92  8 20 56 90 56  8 11  5 12
#> 
#> $BootSamples[[23]]
#>  [1] 81 20 58  4 58 52 68 47 11 90 12 68  4 81 52 92 11 90  8 58 11 90 17 75 43
#> 
#> $BootSamples[[24]]
#>  [1] 81 58 56 45 45 58 11 52 73  5 94 75 12 11  4 52 52 11  4 94 58 58 52 43 17
#> 
#> $BootSamples[[25]]
#>  [1] 11 73 35 35 56 12  8 17 52 81 75  5 45  8 73 90 47 58 43 47 52  8 73 83  8
#> 
#> $BootSamples[[26]]
#>  [1]  8 94 83 90 81 94 35 35 58 83 38 11 94 73 52 45 83 56 92 90  8 58 35 75 68
#> 
#> $BootSamples[[27]]
#>  [1] 83 20 11 17 43 35 35 92 56 73 94 44 11 94 75 47 43 20  4 58 92 47 11 94 83
#> 
#> $BootSamples[[28]]
#>  [1]  8 68 35 38 20 90 68 73 58 20 56 94 90 38 90 52 45 92 43  4 94 45 58 35 43
#> 
#> $BootSamples[[29]]
#>  [1] 52 52 73 38 44 94  5  8 45 12 35 73 11 38 90 83 11 43  5 17 17 44 83 92 75
#> 
#> $BootSamples[[30]]
#>  [1] 35 47  4 73 43 81 58 58 56 92 56 90 56 47 73  5 90 11 90 17 43  8 11  4 35
#> 
#> $BootSamples[[31]]
#>  [1] 43 81 12 45 73 90 38 56 20  4 44  4 75  5 68  4  5 90  4 68 17 44 12 68 52
#> 
#> $BootSamples[[32]]
#>  [1]  8 11 12 35  5 92 90 35 81 17 83 81 38 43 83 58 90 35 73 44  8 38 52 92 92
#> 
#> $BootSamples[[33]]
#>  [1] 12  8 44 83 47 47 35  8 17 73 90 17  4 90 20 17 94  8 58 12 17 43 94 90 45
#> 
#> $BootSamples[[34]]
#>  [1] 11 17 44 11 90 45 52 75 38 17 12 44 56 17 11 58 17 75 94 68  4 17 90 90 81
#> 
#> $BootSamples[[35]]
#>  [1] 12 75 81 47 90 58 44 52 38 94 43 68 90 92 17 17  8  4 58 43 81 56 90 73 94
#> 
#> $BootSamples[[36]]
#>  [1]  4 75 20 75 52 20  8 90 68 81 90 43 20 68 90 38 73 11 90 81 94 92 35 20 90
#> 
#> $BootSamples[[37]]
#>  [1] 81 20 47 83 81 35 81 12 20 38 44 52 20 35 81  4 12 94  8 83  8 56 58 44 35
#> 
#> $BootSamples[[38]]
#>  [1] 52 68 20 44 68 94 94 17 83 47 35 83 47 45 56  5 35 56 38  5 75 90 75 56 20
#> 
#> $BootSamples[[39]]
#>  [1] 17 56 94 81  4 38 58 56  5 83 20 52 20  4  8 45 81 20 94 83 75 44 45 73  4
#> 
#> $BootSamples[[40]]
#>  [1] 73 94 83 11 92 52 35 92 17 75 44 81 90 12 12 75 83 68 90 58 11 35 52 90 52
#> 
#> $BootSamples[[41]]
#>  [1] 12 47 35 58 11  4 20  4 92 73 47  4 11 56 92 35 56  4 94 47 35 11 83 94 38
#> 
#> $BootSamples[[42]]
#>  [1] 12 94 12 73 47 81 45 52 83 94 73 94 35 52 43 90 11 20 92 44  5 90 45 43  8
#> 
#> $BootSamples[[43]]
#>  [1]  8  5 56 17 56 12 47  5 90  4 35 68 56 47 90 38 43 44 73 83 47 47 90 90 73
#> 
#> $BootSamples[[44]]
#>  [1] 56 38  8 56 43 81 90 11 92 43 83 75 56 38 68 90 90 73 68  4  8 35 44 17 44
#> 
#> $BootSamples[[45]]
#>  [1] 47 75 17 83  8 56 47 92  4 12 17 75 20  5 35 47 20 73 38 81 17 35 90 92 43
#> 
#> $BootSamples[[46]]
#>  [1] 35 90 68  4 94  8 73 56 12 43  4 56 44 45 44  5 35 81 43 73 52 43 83  8  5
#> 
#> $BootSamples[[47]]
#>  [1] 81 12 94 68  5 73 45 90 12 35  5 83 92 81 90 68 12 83 20 44 20 17 81 20 90
#> 
#> $BootSamples[[48]]
#>  [1] 94  8 83 47  4  4 44 47 43 90  8 92 11 52 11 44 92  5 75 20 11 52  8 90 43
#> 
#> $BootSamples[[49]]
#>  [1] 90 35 90 75 17 90  8 73 81  8 68 12  4 11 81 45 92 20 52  4  8 58 45 35 52
#> 
#> $BootSamples[[50]]
#>  [1]  8 58 68 11 45 92 45 38  4 11 52 75 56  4 11 90 58 94 17 35 43 58 52 11 81
#> 
#> 
#> $BootEstimates
#>  [1] 40.36 51.80 56.28 54.68 48.32 42.48 47.92 55.64 49.84 51.80 52.52 44.16
#> [13] 52.48 54.16 58.20 51.08 60.56 54.24 52.52 43.00 53.40 47.24 48.04 44.84
#> [25] 43.80 61.56 51.28 54.28 45.60 47.32 40.88 51.84 42.92 45.36 57.00 57.12
#> [37] 45.28 52.32 46.40 59.08 42.52 53.52 48.96 52.44 45.16 44.16 52.84 43.12
#> [49] 46.16 44.68
#> 
#> $CI95
#>     2.5% estimate    97.5% 
#>   41.240   49.640   60.227 
#> 
#> $SE
#> [1] 5.444399
#> 

bootstrap(x = LETTERS[1:26], boot = 50)                      # character

#> [[1]]
#>  [1] "S" "L" "H" "B" "J" "Z" "C" "V" "D" "S" "X" "P" "K" "F" "V" "A" "Z" "O" "P"
#> [20] "H" "I" "R" "A" "L" "L" "M"
#> 
#> [[2]]
#>  [1] "P" "Z" "J" "X" "Y" "X" "Q" "F" "B" "I" "C" "K" "C" "Q" "O" "M" "I" "Q" "J"
#> [20] "K" "P" "O" "Z" "I" "R" "M"
#> 
#> [[3]]
#>  [1] "W" "T" "C" "U" "E" "I" "M" "N" "D" "N" "B" "S" "B" "S" "C" "A" "D" "F" "F"
#> [20] "I" "H" "G" "S" "J" "M" "K"
#> 
#> [[4]]
#>  [1] "L" "O" "X" "W" "C" "N" "Y" "O" "A" "H" "V" "Z" "P" "B" "J" "T" "Q" "X" "B"
#> [20] "Z" "V" "G" "H" "D" "K" "Y"
#> 
#> [[5]]
#>  [1] "W" "S" "L" "Y" "H" "A" "H" "E" "O" "F" "Q" "E" "L" "J" "R" "E" "K" "O" "T"
#> [20] "T" "Z" "V" "H" "S" "R" "W"
#> 
#> [[6]]
#>  [1] "P" "O" "A" "S" "R" "L" "X" "V" "V" "M" "G" "E" "K" "D" "N" "A" "W" "D" "E"
#> [20] "D" "X" "Q" "P" "X" "I" "B"
#> 
#> [[7]]
#>  [1] "S" "L" "O" "E" "K" "W" "S" "Z" "G" "B" "K" "O" "C" "R" "Y" "X" "T" "M" "R"
#> [20] "L" "Q" "I" "V" "A" "G" "X"
#> 
#> [[8]]
#>  [1] "H" "S" "W" "G" "K" "L" "G" "I" "Z" "H" "K" "P" "X" "Y" "I" "B" "S" "S" "R"
#> [20] "B" "S" "Q" "I" "U" "X" "C"
#> 
#> [[9]]
#>  [1] "T" "G" "E" "L" "F" "G" "V" "J" "U" "N" "K" "V" "C" "K" "D" "T" "H" "I" "E"
#> [20] "W" "B" "B" "X" "D" "C" "A"
#> 
#> [[10]]
#>  [1] "V" "J" "L" "B" "N" "L" "X" "F" "W" "P" "P" "R" "T" "E" "B" "W" "X" "K" "G"
#> [20] "Y" "Y" "O" "L" "Y" "B" "K"
#> 
#> [[11]]
#>  [1] "W" "D" "Y" "Q" "R" "N" "K" "Q" "T" "E" "J" "Y" "U" "E" "D" "R" "E" "S" "B"
#> [20] "T" "L" "M" "N" "Z" "O" "E"
#> 
#> [[12]]
#>  [1] "Y" "T" "H" "N" "X" "T" "U" "S" "U" "I" "I" "S" "K" "U" "M" "F" "S" "L" "S"
#> [20] "J" "E" "M" "G" "D" "Q" "V"
#> 
#> [[13]]
#>  [1] "J" "V" "E" "B" "C" "P" "Y" "W" "L" "G" "O" "D" "K" "G" "T" "C" "B" "I" "R"
#> [20] "Y" "G" "H" "X" "M" "G" "Z"
#> 
#> [[14]]
#>  [1] "A" "W" "L" "H" "D" "B" "V" "I" "G" "F" "Z" "T" "R" "T" "S" "T" "K" "H" "L"
#> [20] "C" "D" "G" "H" "U" "R" "N"
#> 
#> [[15]]
#>  [1] "N" "B" "R" "T" "B" "N" "I" "P" "A" "W" "S" "Y" "T" "M" "S" "H" "S" "I" "P"
#> [20] "J" "B" "D" "Z" "E" "O" "G"
#> 
#> [[16]]
#>  [1] "K" "M" "V" "Y" "C" "N" "W" "G" "N" "E" "N" "H" "I" "S" "H" "C" "B" "V" "H"
#> [20] "E" "U" "S" "R" "D" "H" "N"
#> 
#> [[17]]
#>  [1] "U" "Q" "D" "J" "C" "T" "R" "V" "L" "Z" "H" "Y" "G" "N" "J" "Y" "N" "C" "A"
#> [20] "O" "B" "R" "Y" "I" "S" "B"
#> 
#> [[18]]
#>  [1] "D" "X" "G" "W" "H" "H" "O" "H" "U" "L" "W" "H" "S" "K" "Z" "U" "J" "Z" "F"
#> [20] "P" "I" "E" "K" "F" "T" "R"
#> 
#> [[19]]
#>  [1] "U" "G" "M" "E" "U" "D" "D" "E" "Q" "T" "C" "A" "O" "Y" "F" "L" "J" "R" "H"
#> [20] "G" "F" "G" "S" "O" "T" "V"
#> 
#> [[20]]
#>  [1] "W" "C" "D" "M" "I" "L" "Z" "T" "J" "L" "J" "N" "R" "V" "Q" "N" "O" "Q" "O"
#> [20] "E" "S" "E" "N" "J" "X" "J"
#> 
#> [[21]]
#>  [1] "S" "Y" "A" "A" "G" "L" "K" "O" "T" "G" "O" "H" "O" "J" "F" "H" "Y" "J" "K"
#> [20] "Q" "M" "U" "L" "Y" "Y" "P"
#> 
#> [[22]]
#>  [1] "S" "Q" "O" "L" "A" "Q" "B" "J" "A" "F" "K" "M" "R" "L" "P" "L" "P" "T" "B"
#> [20] "P" "P" "A" "X" "N" "L" "I"
#> 
#> [[23]]
#>  [1] "Q" "J" "U" "M" "R" "X" "W" "J" "D" "E" "R" "Y" "K" "F" "U" "P" "U" "E" "L"
#> [20] "D" "H" "D" "A" "K" "M" "X"
#> 
#> [[24]]
#>  [1] "I" "S" "K" "C" "V" "E" "C" "Q" "Y" "R" "L" "G" "A" "A" "O" "P" "J" "K" "C"
#> [20] "R" "H" "S" "Y" "E" "L" "U"
#> 
#> [[25]]
#>  [1] "X" "Y" "M" "S" "D" "Z" "R" "T" "P" "L" "I" "U" "Z" "Z" "T" "J" "H" "Z" "V"
#> [20] "X" "H" "A" "X" "T" "J" "H"
#> 
#> [[26]]
#>  [1] "X" "W" "V" "X" "G" "O" "U" "P" "Y" "V" "A" "M" "G" "G" "J" "E" "U" "K" "P"
#> [20] "B" "B" "M" "G" "T" "P" "S"
#> 
#> [[27]]
#>  [1] "B" "T" "R" "G" "V" "X" "P" "D" "V" "T" "S" "H" "K" "J" "A" "W" "N" "Y" "M"
#> [20] "W" "Y" "Q" "K" "A" "F" "I"
#> 
#> [[28]]
#>  [1] "N" "M" "J" "R" "C" "C" "R" "R" "G" "D" "L" "R" "R" "K" "Q" "I" "U" "D" "I"
#> [20] "V" "W" "K" "Y" "U" "P" "X"
#> 
#> [[29]]
#>  [1] "Q" "G" "X" "J" "Y" "D" "I" "V" "Q" "I" "D" "D" "W" "P" "P" "Y" "C" "S" "X"
#> [20] "F" "L" "O" "I" "A" "E" "X"
#> 
#> [[30]]
#>  [1] "E" "Q" "P" "M" "S" "B" "N" "J" "U" "T" "G" "R" "A" "L" "P" "K" "C" "T" "B"
#> [20] "M" "J" "D" "Q" "S" "Y" "F"
#> 
#> [[31]]
#>  [1] "K" "U" "M" "W" "W" "X" "Q" "R" "Z" "C" "E" "U" "C" "G" "K" "J" "X" "G" "T"
#> [20] "D" "A" "H" "F" "C" "L" "G"
#> 
#> [[32]]
#>  [1] "S" "A" "S" "Q" "Y" "S" "X" "I" "V" "E" "W" "D" "H" "X" "K" "H" "V" "G" "S"
#> [20] "C" "P" "U" "H" "Z" "O" "W"
#> 
#> [[33]]
#>  [1] "Q" "A" "Y" "Y" "Z" "S" "Z" "S" "Z" "G" "O" "D" "I" "T" "E" "R" "N" "B" "R"
#> [20] "C" "E" "X" "Z" "S" "C" "I"
#> 
#> [[34]]
#>  [1] "D" "T" "T" "M" "A" "C" "K" "T" "D" "T" "W" "Z" "G" "A" "L" "M" "S" "G" "M"
#> [20] "J" "D" "T" "T" "O" "F" "V"
#> 
#> [[35]]
#>  [1] "I" "J" "C" "K" "Z" "K" "E" "J" "C" "Z" "E" "A" "Z" "T" "Z" "J" "V" "Q" "T"
#> [20] "J" "Q" "N" "A" "W" "H" "I"
#> 
#> [[36]]
#>  [1] "U" "V" "D" "G" "Q" "U" "L" "R" "M" "G" "Z" "J" "G" "D" "U" "S" "X" "O" "P"
#> [20] "R" "U" "A" "F" "W" "H" "E"
#> 
#> [[37]]
#>  [1] "U" "H" "F" "N" "J" "C" "W" "B" "A" "W" "L" "V" "O" "G" "Q" "L" "X" "V" "V"
#> [20] "R" "Q" "Q" "O" "B" "L" "G"
#> 
#> [[38]]
#>  [1] "W" "B" "V" "U" "X" "Z" "W" "F" "O" "T" "Q" "S" "G" "Q" "D" "E" "C" "X" "H"
#> [20] "P" "O" "X" "H" "K" "K" "R"
#> 
#> [[39]]
#>  [1] "C" "D" "P" "D" "W" "Z" "G" "I" "C" "N" "K" "H" "M" "L" "X" "Z" "U" "F" "O"
#> [20] "J" "W" "T" "E" "B" "J" "I"
#> 
#> [[40]]
#>  [1] "W" "M" "O" "V" "D" "K" "V" "E" "S" "B" "E" "Z" "X" "D" "L" "T" "O" "X" "Z"
#> [20] "Q" "T" "K" "B" "N" "Z" "J"
#> 
#> [[41]]
#>  [1] "J" "Y" "Y" "N" "G" "Q" "K" "F" "S" "R" "V" "G" "R" "D" "E" "U" "J" "Y" "T"
#> [20] "O" "O" "Q" "M" "X" "K" "D"
#> 
#> [[42]]
#>  [1] "P" "Z" "Q" "A" "V" "H" "E" "K" "T" "Y" "G" "R" "X" "U" "F" "D" "N" "V" "Y"
#> [20] "J" "V" "A" "U" "X" "L" "Z"
#> 
#> [[43]]
#>  [1] "F" "C" "W" "Z" "L" "U" "D" "T" "A" "Y" "B" "C" "L" "N" "V" "P" "C" "E" "Q"
#> [20] "S" "F" "W" "T" "C" "Y" "J"
#> 
#> [[44]]
#>  [1] "I" "A" "N" "N" "Y" "T" "E" "L" "Z" "D" "L" "E" "H" "G" "L" "P" "X" "I" "K"
#> [20] "Y" "S" "U" "C" "E" "C" "K"
#> 
#> [[45]]
#>  [1] "K" "U" "C" "M" "X" "O" "G" "J" "U" "B" "N" "F" "L" "F" "Y" "P" "X" "Q" "N"
#> [20] "K" "Q" "F" "D" "P" "P" "D"
#> 
#> [[46]]
#>  [1] "J" "O" "I" "G" "T" "H" "N" "C" "A" "X" "M" "R" "N" "L" "J" "T" "V" "R" "K"
#> [20] "T" "G" "U" "Q" "L" "G" "W"
#> 
#> [[47]]
#>  [1] "A" "M" "W" "W" "E" "M" "C" "F" "L" "U" "R" "X" "W" "R" "C" "T" "W" "A" "L"
#> [20] "H" "L" "W" "K" "H" "Z" "T"
#> 
#> [[48]]
#>  [1] "Q" "G" "X" "Y" "V" "G" "V" "C" "X" "Q" "B" "E" "O" "J" "A" "H" "W" "Y" "U"
#> [20] "S" "V" "N" "Q" "Q" "P" "X"
#> 
#> [[49]]
#>  [1] "W" "P" "I" "O" "A" "T" "R" "Q" "D" "R" "U" "Q" "T" "Y" "H" "S" "H" "C" "N"
#> [20] "K" "J" "J" "R" "B" "K" "O"
#> 
#> [[50]]
#>  [1] "R" "Q" "N" "O" "J" "C" "R" "K" "W" "H" "S" "Y" "N" "B" "S" "K" "Q" "B" "O"
#> [20] "W" "V" "C" "O" "K" "Q" "H"
#> 

z <- factor(sample(c("stu", "is", "cool"), 10, replace = TRUE))
# bootstrap(x = z, boot = 50)      # factor (ask to convert to character)