Linear Sensitivity Analysis

Sensitivity analysis of linear maps, all in one function.

calcMatrixSensitivity(A)

Arguments

A
A projection matrix to be analyzed.

Value

A list of the following:

Matrix

The original projection matrix (map)

Sensitivity

Matrix of linear sensitivity based on perturbations of parameters to lambda

Elasticity

Matrix of linear elasticity (proportional change) based on perturbations of parameters to lambda

lambda

The dominant eigenvalue of A. Projection matrix population "growth"

w

The right eigenvector or stable stage distribution of the population

v

The left eigenvector or the reproductive "value" of each stage in the population

References

Caswell, H. Matrix Population Models. 2001. Sensitivity & Elasticity.

See also

Examples

A <- diag(1:5 / 10)
A[cbind(2:5, 1:4)] <- 3:6 / 10
A[1, 5] <- 5

calcMatrixSensitivity(A)

#> $Matrix
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  0.1  0.0  0.0  0.0  5.0
#> [2,]  0.3  0.2  0.0  0.0  0.0
#> [3,]  0.0  0.4  0.3  0.0  0.0
#> [4,]  0.0  0.0  0.5  0.4  0.0
#> [5,]  0.0  0.0  0.0  0.6  0.5
#> 
#> $Sensitivity
#>           [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 0.1505515 0.05482484 0.03029776 0.02428429 0.02781632
#> [2,] 0.4636053 0.16882649 0.09329829 0.07478055 0.08565701
#> [3,] 0.9548113 0.34770401 0.19215106 0.15401314 0.17641358
#> [4,] 1.3822113 0.50334597 0.27816321 0.22295369 0.25538120
#> [5,] 1.4370710 0.52332366 0.28920344 0.23180267 0.26551722
#> 
#> $Elasticity
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] 0.01470497 0.00000000 0.00000000 0.00000000 0.1358466
#> [2,] 0.13584657 0.03297992 0.00000000 0.00000000 0.0000000
#> [3,] 0.00000000 0.13584657 0.05630449 0.00000000 0.0000000
#> [4,] 0.00000000 0.00000000 0.13584657 0.08710712 0.0000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.13584657 0.1296706
#> 
#> $lambda
#> [1] 1.023814
#> 
#> $w
#> [1] 0.9004743 0.3279167 0.1812160 0.1452485 0.1663741
#> 
#> $v
#> [1] 0.06650293 0.20478775 0.42176748 0.61056231 0.63479540
#>