Calculate Interclass Correlation Coefficient (ICC)

The Intraclass correlation coefficient (r_i; aka ICC) can be used to estimate the repeatability of a method. The value 0 -> 1. Depending on how the groups are set up, you want all your variation to be among groups (individuals), not within groups (repeats) so you want this value to be high if individuals are your groups, and low if your repeated measurements are the groups. When ICC is high, it means most of the variation is between treatment groups.

coefRi(x, groups, do.log = TRUE)

Arguments

x
A matrix or data frame containing the raw data of the various treatments you are testing in the ANOVA

groups
A vector of the factor groupings for x

do.log
Should data be performed on log10-transformed?

Value

A list containing:

model

Resulting ANOVA table

ICC

The intraclass correlation coefficient, the measure of similarity among individuals within a treatment group relative to the differences found among groups

References

Sokal & Rohlf (Biometry; 3rd ed.) 210-214. Sokal & Rohlf (Biometry; 2rd ed.) 211-216.

See also

Examples

head(Ri_data)    # internal data

#>   one two three four
#> 1 380 350   354  376
#> 2 376 356   360  344
#> 3 360 358   362  342
#> 4 368 376   352  372
#> 5 372 338   366  374
#> 6 366 342   372  360

x <- as.vector(as.matrix(Ri_data))[ !is.na(as.vector(as.matrix(Ri_data))) ]
coefRi(x, groups = rep(names(Ri_data), c(8, 10, 13, 6)))

#> $Model
#>             Df   Sum Sq   Mean Sq F value  Pr(>F)   
#> groups       3 0.002603 0.0008677   5.158 0.00492 **
#> Residuals   33 0.005552 0.0001682                   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> $ICC
#> [1] 0.3157723
#>