Calculate Limited Cumulative Sum

Calculate the cumulative sum of a set of numbers within a vector. The difference between this and cumsum is that it is a sliding window approach, so sums are not necessarily calculated over the entire length of the vector. When cut == length(x) then it is the same as cumsum. Also, for entries < window, i.e. the beginning, the entries returned will be identical to cumsum.

cumsumWindow(x, window)

Arguments

x
The vector to be summed across.

window
The length/size of the window to sum within (the moving cutoff).

Value

A vector of the sums of the sliding window for the cumulative sums.

Details

If window >= length(x), a warning is triggered and cumsumWindow reverts to cumsum.

See also

Examples

cumsumWindow(1:20, 5)

#>  [1]  1  3  6 10 15 20 25 30 35 40 45 50 55 60 65 70 75 80 85 90

cumsumWindow(1:20, window = 20)

#> Warning: window >= length(x)!

#>  [1]   1   3   6  10  15  21  28  36  45  55  66  78  91 105 120 136 153 171 190
#> [20] 210

cumsum(1:20)

#>  [1]   1   3   6  10  15  21  28  36  45  55  66  78  91 105 120 136 153 171 190
#> [20] 210

r.vec <- sample(1:20, 100, replace=TRUE) # random vector
cumsumWindow(r.vec, 5)

#>   [1] 19 31 47 63 72 69 61 53 53 45 43 58 70 61 69 64 49 46 47 39 40 51 43 49 53
#>  [26] 56 58 57 47 56 47 46 46 44 39 38 28 33 32 35 36 48 55 62 69 82 73 69 71 65
#>  [51] 59 62 66 61 67 72 79 75 87 74 74 67 55 40 39 29 35 34 37 41 41 26 26 40 50
#>  [76] 55 65 65 49 48 40 43 59 72 55 56 44 30 27 42 47 60 56 52 53 63 59 74 85 76