Map Vector

Map a factor or character vector (1-to-1 or many-2-one) to another new vector.

mapVector(v, from.list, to.vec, nomatch = NULL)

Arguments

v
The source vector to perform the mapping. Must be a character or factor vector, or be able to be converted to one. In general, the returned value is of the same class (character/factor) as the input vector.

from.list
A list of the associated levels to map in the order they will be mapped to in to.vec. Alternatively, a character vector of the same length as to.vec of only 1-to-1 mapping is desired.

to.vec
The new mapping in the same order as they appear in from.list.

nomatch
Value that unmapped values in the source vector should take. By default, NULL, unmapped values are themselves returned unchanged.

...
Additional arguments passed to internals of specific S3 methods.

Value

A factor (or character) vector with the new mapping applied.

Details

If v is a factor, the returned vector is also a factor (and the levels of the returned value are determined by the to.vec argument), otherwise class of the returned vector is determined by the to.vec argument. If the class of the returned vector changes, a warning is flagged.

factor -> factor character -> factor, numeric, or integer

See also

Examples

# factor
mapVector(test_data$TimePoint,
          list(c("baseline", "24 months"), c("6 months", "12 months")),
          c("one", "two"))

#>  [1] one one one one one one one one one one two two two two two two two two two
#> [20] two two two two two two two two two two two one one one one one one one one
#> [39] one one
#> Levels: one two

# numeric -> character (with warning)
mapVector(rep(1:4, each = 10), list(c(1, 2), c(3, 4)), c("one", "two"))

#> Warning: * mapVector is changing object class of [v]: integer -> character

#>  [1] "one" "one" "one" "one" "one" "one" "one" "one" "one" "one" "one" "one"
#> [13] "one" "one" "one" "one" "one" "one" "one" "one" "two" "two" "two" "two"
#> [25] "two" "two" "two" "two" "two" "two" "two" "two" "two" "two" "two" "two"
#> [37] "two" "two" "two" "two"

# character ("C" is un-mapped)
mapVector(head(LETTERS, 5), list(c("A", "B"), c("D", "E")), c("one", "two"))

#> [1] "one" "one" "C"   "two" "two"

# character with missing (nomatch)
mapVector(head(LETTERS, 5), list(c("A", "B"), c("D", "E")), c("one", "two"), nomatch = "No")

#> [1] "one" "one" "No"  "two" "two"

# integer
mapVector(1:5, list(c(1, 3, 5), c(2, 4)), c(10, 100))

#> Warning: * mapVector is changing object class of [v]: integer -> numeric

#> [1] "10"  "100" "10"  "100" "10" 

# character -> integer
mapVector(head(LETTERS, 5), list(c("A", "B"), c("C", "D", "E")), 1:2)

#> Warning: * mapVector is changing object class of [v]: character -> integer

#> [1] 1 1 2 2 2