Merge Meta Data

Merge additional meta data to an existing data frame based on unique row sample identifiers. This is a wrapper around the existing left_join.

mergeMetaData(df, meta, ...)

Arguments

df
The existing data frame to which the meta data is to be merged.

meta
Can be either:

A data frame of additional meta data, with at least one common column name to match the samples OR

A file name (and path) pointing to where to get a "*.csv" file containing meta data (this option uses a call to read.csv).

...
Additional arguments passed to left_join.

Value

A tibble object with extra columns corresponding to the meta data added and indexed to the appropriate sample IDs.

See also

Examples

set.seed(101)
new <- data.frame(pid = sample(test_data$pid, 10), new.clin = rnorm(10))
mergeMetaData(test_data, new)            # use default 'by'

#> Joining, by = "pid"

#> # A tibble: 40 x 13
#>      pid Pop   Sample TimePoint     a     b ABCD.1234.56.8 XYZZ.6969.4.7      x
#>    <int> <chr> <fct>  <fct>     <dbl> <int>          <dbl>         <dbl>  <dbl>
#>  1  1041 A     medium baseline    157     2           31.5          24.0  5.11 
#>  2  1042 B     small  baseline    174     0           28.9          23.8  6.18 
#>  3  1043 C     large  baseline    155     0           23.2          22.0  0.488
#>  4  1044 D     medium baseline    131     3           23.1          24.2  0.239
#>  5  1045 E     small  baseline    175     1           18.9          25.7  2.22 
#>  6  1046 F     small  baseline    132     5           26.6          23.8  7.33 
#>  7  1047 G     medium baseline    169     3           25.0          23.9  2.66 
#>  8  1048 H     medium baseline    158     3           29.7          24.6  1.16 
#>  9  1049 I     medium baseline    169     4           27.5          27.5  5.00 
#> 10  1050 J     medium baseline    170     0           30.4          23.1 -0.553
#> # … with 30 more rows, and 4 more variables: y <dbl>, z <dbl>, Response <fct>,
#> #   new.clin <dbl>

mergeMetaData(test_data, new, by = "pid") # same

#> # A tibble: 40 x 13
#>      pid Pop   Sample TimePoint     a     b ABCD.1234.56.8 XYZZ.6969.4.7      x
#>    <int> <chr> <fct>  <fct>     <dbl> <int>          <dbl>         <dbl>  <dbl>
#>  1  1041 A     medium baseline    157     2           31.5          24.0  5.11 
#>  2  1042 B     small  baseline    174     0           28.9          23.8  6.18 
#>  3  1043 C     large  baseline    155     0           23.2          22.0  0.488
#>  4  1044 D     medium baseline    131     3           23.1          24.2  0.239
#>  5  1045 E     small  baseline    175     1           18.9          25.7  2.22 
#>  6  1046 F     small  baseline    132     5           26.6          23.8  7.33 
#>  7  1047 G     medium baseline    169     3           25.0          23.9  2.66 
#>  8  1048 H     medium baseline    158     3           29.7          24.6  1.16 
#>  9  1049 I     medium baseline    169     4           27.5          27.5  5.00 
#> 10  1050 J     medium baseline    170     0           30.4          23.1 -0.553
#> # … with 30 more rows, and 4 more variables: y <dbl>, z <dbl>, Response <fct>,
#> #   new.clin <dbl>