Determine Path to File

Determine the path of a defined file via brute force search of root directory and optionally change the R working directory to the directory containing that file. Useful in setting the working directory prior to a read.csv call.

path2file(file, root.dir = Sys.getenv("HOME"), switchDir = FALSE)

Arguments

file
Character string of the desired file, typically "*.R" or "*.csv".

root.dir
The path of the root directory. The higher up the root to travel the longer the search will take If you narrow down the path, more specific, it will take less time. The default is a Mac/Unix formulation. For Windows/PC machines, root.dir must be specified.

switchDir
Logical. Set the working directory to the path of the file in file, if found. Uses setwd.

Value

The path of the file specified in file.

Details

If the file in filename is not found, function is stopped and warning printed.

References

Assistance from Steven Mosher via StackOverflow

See also

Examples

if (FALSE) {
filename <- "TreeData.csv"
path2file(filename)
path2file(filename, switchDir = TRUE)
read.csv(filename)
}