Print Object to File

A convenient wrapper around the sink function to print the resulting print command to a corresponding file connection. Useful for sharing data only results with collaborators or medical group.

print2file(..., file, width = 250)

Arguments

...
Object(s) to be printed to file as standard output.

file
File name for the output.

width
Numeric. Controls the maximum number of columns on a line used in printing vectors, matrices and arrays, and when filling by cat. See also getOption("width").

See also

Examples

tab <- crossTab(test_data, c("Pop", "Sample"))
print2file(tab, file = "table_file.txt")

#> * Creating text file ... table_file.txt