Search & Replace

A global search & replace of entries within a vector, matrix, or data frame Note: The lengths of s and r must be identical.

searchReplace(x, s, r)

Arguments

x
The object to be searched, typically a matrix or data frame but can be a vector of character or numeric class.

s
The search index.

r
The replace with. Must be same length as s.

...
Pass through to various S3 methods.

Value

An object of the same dimensions and class as x, with the s= matches replaced with r=.

See also

Examples

# matrix
Y <- matrix(1:25, ncol = 5)
Y

#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    6   11   16   21
#> [2,]    2    7   12   17   22
#> [3,]    3    8   13   18   23
#> [4,]    4    9   14   19   24
#> [5,]    5   10   15   20   25

searchReplace(Y, s = c(8, 20), r = c(99, 99))

#> Replacing: 8 -> 99 
#> Replacing: 20 -> 99 

#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    6   11   16   21
#> [2,]    2    7   12   17   22
#> [3,]    3   99   13   18   23
#> [4,]    4    9   14   19   24
#> [5,]    5   10   15   99   25

# data.frame
X <- data.frame(x = c(1, 2, 3), y = c(2, 2, 4), z = c(1, 2, 4))
rownames(X) <- c("one", "two", "three")
X

#>       x y z
#> one   1 2 1
#> two   2 2 2
#> three 3 4 4

searchReplace(X, s = 1:4, r = c("A", "C", "G", "T"))

#> Replacing: 1 -> A 
#> Replacing: 2 -> C 
#> Replacing: 3 -> G 
#> Replacing: 4 -> T 

#>       x y z
#> one   A C A
#> two   C C C
#> three G T T

# numeric
searchReplace(1:10, s = 4, r = 19)

#> Replacing: 4 -> 19 

#>  [1]  1  2  3 19  5  6  7  8  9 10

# character
searchReplace(head(LETTERS, 10), s = "G", r = "Z")

#> Replacing: G -> Z 

#>  [1] "A" "B" "C" "D" "E" "F" "Z" "H" "I" "J"