Apply Function to Subsets of Data

Apply a function, written on the fly (.fun), to the subsets of data determined by an appropriate subsetting variable (column) If the x is a vector, the index argument must be of the same length as x, and the function reduces to a simple tapply. If x is a data frame, then y must be supplied and corresponds to a column in x.

subapply(x, index, .fun, ...)

# S3 method for data.frame
subapply(x, index, .fun, y, longform = FALSE,
  sum.field = deparse(substitute(.fun)), ...)

Arguments

x
A vector, data frame, of matrix of values evaluate.

index
If x is a vector, the sub-setting index of same length as x. If x is a data frame, a character string indicating the column name(s) in x to use as indices. Character string currently cannot be more than length 2.

.fun
The function to apply to each subset of the data. Can be written on the fly using the function(x) syntax.

...
Additional arguments passed to tapply

y
Character. Only used if x is a data frame, then a character string corresponding to a column name in x.

longform
Alternative version of output for the returned value. Returns a table.

sum.field
Character. The summary field column name in the returned data frame that summarizes the function called upon each subset of the data. Ignored if longform = FALSE.

Value

A data frame with summary data, and the result of the applied function.

Details

THIS FUNCTION WILL EVENTUALLY BE REPLACED BY summarise.

See also

Examples

# S3 numeric
subapply(test_data$z, index = test_data$Sample, .fun = mean)

#>     small    medium     large 
#> 0.7525833 1.1068000 1.0505385 

# S3 character
set.seed(10)
y <- sample(LETTERS, 10, replace = TRUE)
subapply(y, rep("a",length(y)), duplicated)

#> $a
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE  TRUE
#> 

# S3 data.frame
subapply(test_data, index = "Sample", .fun = mean, y = "z")  # same as above

#> Sample
#>     small    medium     large 
#> 0.7525833 1.1068000 1.0505385 

# cross tabulated
subapply(test_data, c("Sample", "Response"), .fun = mean, y = "z")

#>         Response
#> Sample     Control   Disease
#>   small  0.6418333 0.8633333
#>   medium 1.1521818 0.9820000
#>   large  1.1673333 1.0155000

# cross tabulated long-format
subapply(test_data, index = c("Sample","Response"), .fun = mean, y = "z",
         longform = TRUE, sum.field = "Mean")

#>   c(2, 1, 3, 2, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 2, 2, 3, 1, 2, 1, 3, 3, 1, 2, 3, 3, 1, 2, 2, 2, 2, 1, 1, 1, 2, 3, 1, 3, 1, 3)
#> 1                                                                                                                     small
#> 2                                                                                                                    medium
#> 3                                                                                                                     large
#> 4                                                                                                                     small
#> 5                                                                                                                    medium
#> 6                                                                                                                     large
#>   c(1, 2, 1, 1, 2, 2, 1, 1, 2, 2, 2, 2, 2, 1, 1, 1, 2, 1, 1, 1, 2, 2, 1, 1, 1, 2, 2, 2, 1, 1, 2, 2, 2, 1, 1, 2, 1, 2, 1, 2)
#> 1                                                                                                                   Control
#> 2                                                                                                                   Control
#> 3                                                                                                                   Control
#> 4                                                                                                                   Disease
#> 5                                                                                                                   Disease
#> 6                                                                                                                   Disease
#>    n      Mean
#> 1  6 0.6418333
#> 2 11 1.1521818
#> 3  3 1.1673333
#> 4  6 0.8633333
#> 5  4 0.9820000
#> 6 10 1.0155000

# S3 matrix
test_data$Sample %<>% as.numeric
m <- test_data[, c(1, 3, 11)] %>% data.matrix()
subapply(m, index = "Sample", .fun = mean, y = "z", longform = TRUE) # same as data.frame

#>    n      mean
#> 1 12 0.7525833
#> 2 15 1.1068000
#> 3 13 1.0505385